+++ src/components/tasks/quick/TaskQuickAdd.js
@@ -1,11 +1,13 @@
 import React, { useRef, useState } from 'react';
 import { Select } from 'antd';
+import moment from 'moment';
 import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
 import { ContextTitle } from 'components/contexts/ContextTitle';
 import { FolderTitle } from 'components/folders/FolderTitle';
 import { GoalTitle } from 'components/goals/GoalTitle';
 import { LocationTitle } from 'components/locations/LocationTitle';
+import TagsTitle from 'components/tags/TagsTitle';
 import { TaskTemplateTitle } from 'components/tasktemplates/TaskTemplateTitle';
 import withBusyCheck from 'containers/WithBusyCheck';
 import { useAppApi } from 'hooks/UseAppApi';
@@ -13,16 +15,17 @@ import { useContextApi } from 'hooks/UseContextApi';
 import { useFolderApi } from 'hooks/UseFolderApi';
 import { useGoalApi } from 'hooks/UseGoalApi';
 import { useLocationApi } from 'hooks/UseLocationApi';
+import { useSettingsApi } from 'hooks/UseSettingsApi';
+import { useTagApi } from 'hooks/UseTagApi';
 import { useTaskTemplateApi } from 'hooks/UseTaskTemplateApi';
 import { useTaskApi } from 'hooks/UseTaskApi';
 import { useTaskFieldApi } from 'hooks/UseTaskFieldApi';
 import { applyTaskTemplate, applyTaskTemplateFromTaskFilter } from 'utils/TemplateUtils';
 
 function TaskQuickAdd({ apis }) {
-    const { appApi, contextApi, folderApi, goalApi, locationApi, taskApi, taskFieldApi, taskTemplateApi } = apis;
+    const { appApi, contextApi, folderApi, goalApi, locationApi, settingsApi, tagApi, taskApi, taskFieldApi, taskTemplateApi } = apis;
 
     const [values, setValues] = useState([]);
-    const [open, setOpen] = useState(false);
     const selectRef = useRef(null);
 
     const onChange = values => {
@@ -31,20 +34,6 @@ function TaskQuickAdd({ apis }) {
         } else {
             setValues(values);
         }
-    };
-
-    const onKeyInputDown = () => {
-        setOpen(true);
-    };
-
-    const onFocus = () => {
-        if (values.length > 0) {
-            setOpen(true);
-        }
-    };
-
-    const onBlur = () => {
-        setOpen(false);
     };
 
     const onAdd = async values => {
@@ -56,17 +45,43 @@ function TaskQuickAdd({ apis }) {
         applyTaskTemplateFromTaskFilter(taskApi.selectedTaskFilter, taskTemplateApi.taskTemplates, newTask, taskFieldApi.taskFields);
 
         values.forEach((value, index) => {
+            value = value.trim();
+
+            if (!value) {
+                return;
+            }
+
             if (index === 0) {
                 return;
             }
 
-            const object = JSON.parse(value.substr(value.lastIndexOf('__') + 2));
+            let object = null;
 
-            if (object.field === 'taskTemplate') {
-                const taskTemplate = taskTemplateApi.taskTemplates.find(taskTemplate => taskTemplate.id === object.value);
-                applyTaskTemplate(taskTemplate, newTask, taskFieldApi.taskFields);
-            } else {
-                newTask[object.field] = object.value;
+            try {
+                if (value.includes('__')) {
+                    const o = JSON.parse(value.substr(value.lastIndexOf('__') + 2));
+
+                    if (typeof o === 'object' && o.field && Object.prototype.hasOwnProperty.call(o, 'value')) {
+                        object = o;
+                    }
+                }
+            } catch (e) {
+                // Ignore
+            }
+
+            if (object) {
+                if (object.field === 'taskTemplate') {
+                    const taskTemplate = taskTemplateApi.taskTemplates.find(taskTemplate => taskTemplate.id === object.value);
+                    applyTaskTemplate(taskTemplate, newTask, taskFieldApi.taskFields);
+                } else if (object.field === 'tags') {
+                    newTask.tags = [...(newTask.tags || []), object.value];
+                } else if (object.field === 'startDate') {
+                    newTask.startDate = moment().add(object.value.amount, object.value.unit).toISOString();
+                } else if (object.field === 'dueDate') {
+                    newTask.dueDate = moment().add(object.value.amount, object.value.unit).toISOString();
+                } else {
+                    newTask[object.field] = object.value;
+                }
             }
         });
 
@@ -76,7 +91,6 @@ function TaskQuickAdd({ apis }) {
         appApi.setEditingCell(task.id, 'title');
 
         setValues([]);
-        setTimeout(() => setOpen(false));
     };
 
     return (
@@ -86,15 +100,14 @@ function TaskQuickAdd({ apis }) {
             value={values}
             placeholder="Quick add task..."
             onChange={onChange}
-            onInputKeyDown={onKeyInputDown}
-            onFocus={onFocus}
-            onBlur={onBlur}
-            open={open}
             className="joyride-task-quick-add"
             style={{ width: '100%', padding: 3 }}>
             {values.length > 0 ? [
                 <Select.Option key='add' value="__ADD__">
                     <Icon icon="plus" text="Create task" />
+                </Select.Option>,
+                <Select.Option key="star" value={'Star__' + JSON.stringify({ field: 'star', value: true })}>
+                    <Icon icon="star" text="Star" color="#fcde35" />
                 </Select.Option>,
                 <Select.OptGroup key='contexts' label="Contexts">
                     {contextApi.contexts.map(context => (
@@ -124,10 +137,39 @@ function TaskQuickAdd({ apis }) {
                         </Select.Option>
                     ))}
                 </Select.OptGroup>,
+                <Select.OptGroup key='startDates' label="Start Dates">
+                    {settingsApi.settings.postponeTimeDurations.map(timeDuration => {
+                        const title = `Start date in ${timeDuration.amount} ${timeDuration.unit}${timeDuration.amount > 1 ? 's' : ''}`;
+
+                        return (
+                            <Select.Option key={'startDate_' + timeDuration.id} value={title + '__' + JSON.stringify({ field: 'startDate', value: timeDuration })}>
+                                <Icon icon="calendar-alt" text={title} />
+                            </Select.Option>
+                        );
+                    })}
+                </Select.OptGroup>,
+                <Select.OptGroup key='dueDates' label="Due Dates">
+                    {settingsApi.settings.postponeTimeDurations.map(timeDuration => {
+                        const title = `Due date in ${timeDuration.amount} ${timeDuration.unit}${timeDuration.amount > 1 ? 's' : ''}`;
+
+                        return (
+                            <Select.Option key={'dueDate_' + timeDuration.id} value={title + '__' + JSON.stringify({ field: 'dueDate', value: timeDuration })}>
+                                <Icon icon="calendar-alt" text={title} />
+                            </Select.Option>
+                        );
+                    })}
+                </Select.OptGroup>,
                 <Select.OptGroup key='taskTemplates' label="Task Templates">
                     {taskTemplateApi.taskTemplates.map(taskTemplate => (
                         <Select.Option key={taskTemplate.id} value={taskTemplate.title + '__' + JSON.stringify({ field: 'taskTemplate', value: taskTemplate.id })}>
                             <TaskTemplateTitle taskTemplateId={taskTemplate.id} />
+                        </Select.Option>
+                    ))}
+                </Select.OptGroup>,
+                <Select.OptGroup key='tags' label="Tags">
+                    {tagApi.tags.map(tag => (
+                        <Select.Option key={tag.id} value={tag.title + '__' + JSON.stringify({ field: 'tags', value: tag.id })}>
+                            <TagsTitle tagIds={[tag.id]} />
                         </Select.Option>
                     ))}
                 </Select.OptGroup>
@@ -146,6 +188,8 @@ export default withBusyCheck(TaskQuickAdd, () => ({
     folderApi: useFolderApi(),
     goalApi: useGoalApi(),
     locationApi: useLocationApi(),
+    settingsApi: useSettingsApi(),
+    tagApi: useTagApi(),
     taskApi: useTaskApi(),
     taskFieldApi: useTaskFieldApi(),
     taskTemplateApi: useTaskTemplateApi()