+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New Features
 
 * Add week list calendar view (#115)
+* Improve "Quick Add Task" action (#129)
 * Add setting to customize postpone menu items (#130)
 * Set data folder location via an environment variable (TASKUNIFIER_DATA_FOLDER) (#196)
 * New logging and debug tool